net/mlx5e: Increase flow flags size
Change flow flags type from u8 to u32, so we have more
room for additional upcoming flags.

Issue: 1323489

[juergh: IBM patchset mlx-4.15.0-24.26-patches-20180921 (SF #00199436).]

+++ drivers/net/ethernet/mellanox/mlx5/core/en_tc.c
@@ -2749,7 +2749,7 @@ err_free:
 static int
 __mlx5e_add_fdb_flow(struct mlx5e_priv *priv,
 		     struct tc_cls_flower_offload *f,
-		     u8 flow_flags,
+		     u32 flow_flags,
 		     struct mlx5_eswitch_rep *in_rep,
 		     struct mlx5_core_dev *in_mdev,
 		     struct mlx5e_tc_flow **__flow)
@@ -2807,7 +2807,7 @@ out:
 static int
 mlx5e_add_fdb_flow(struct mlx5e_priv *priv,
 		   struct tc_cls_flower_offload *f,
-		   u8 flow_flags,
+		   u32 flow_flags,
 		   struct mlx5e_tc_flow **__flow)
 {
 	struct mlx5_eswitch *esw = priv->mdev->priv.eswitch;
